backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.40.45",
    "fastapi>=0.118.0",
    "google-genai>=1.41.0",
    "mangum>=0.19.0",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
